Fix minimist vulnerability
Dependabot was not able to do this. It seems that old versions of `mkdirp` are the issue, but updating that causes issues with some packages because its API has changed since then. However, those issues don't seem to occur with this resolution directive. I tested this by deleting the  folder and then yarn install v1.22.1 `yarn install` to verify that the install completes.