Change request info logging to not write request headers.
We are currently getting errors from Google due to our log exporting
creating too many bigquery column names. This is causing us to lose
some logs exports. The issue is that the nested headers metadata end
up being converting to one bigquery column per unique column name. Due
to there being so many possible incoming request header names we are
exceeding the max number of columns.

The relevant log exporter is `winston_logs_to_bigquery`
https://console.cloud.google.com/logs/router?project=khan-academy

The solution implemented here is to not log the request headers at the
INFO level, but to do another log call at the DEBUG level that
includes all the headers. The result will be that there are two logs
written to stackdriver for each requst. But there will be less data
exported to bigquery since the log exporter only sends INFO and
greater logs to bigquery. We should still be able to the headers in
the stackdriver log viewer but in the DEBUG log lines, not the INFO
ones.

Issue: none

Test plan:
- yarn test
- I am not sure how to test this manually.