Rename headers to allHeaders.
If we leave the headers name the same I think that we'll run into an
issue in the log exporter where the type of the
`jsonPayload.metadata.headers` column in the existing bigquery table
(record) will differ from the new type (string). This will cause log
exporting problems for a day.

Instead we'll change the name of the column.

Test plan:
yarn test