[FEI-2722] Add new cache tracking approach (#465)
## Summary:
This removes the old cache tracking, which only worked for in-memory
cache, and replaces it with something more robust.

A related change will be needed in the webapp usage when this new
version is incorporated since this changes the API.

The way it works now is that the calling code includes a callback in
the request options that returns a cacheID for the request about to be
made. That cacheID is then attached to any fresh response. Therefore, it
can then be used to determine if the response was from the cache or not,
as long as the cacheID is generated in a manner that is unique for a
given batch of requests.

For webapp, we can use the requestID from GAE as our cacheID, therefore
allowing us to see for each request, what was from cache and what was
not, just as we did before.

Issue: FEI-2722

## Test plan:
`yarn test`

Add this to webapp and implement to use requestID as cacheID, then use the
render-gateway and observe that the source label on requests for JS files says
"new request" on the first request, but "cache" for subsequent requests.

Author: somewhatabstract

Reviewers: jeresig

Required Reviewers: 

Approved by: jeresig

Checks: ✅ Test (macOS-latest, 12.x), ✅ Test (macOS-latest, 10.x), ✅ Test (ubuntu-latest, 10.x), ✅ Coverage (ubuntu-latest, 12.x), ✅ Lint and flow check (ubuntu-latest, 12.x), ✅ Publish to branch, ⏭  Auto-approve dependabot PRs

Pull request URL: https://github.com/Khan/render-gateway/pull/465