[FEI-2726] Aborted requests to give no response
Superagent will return a partial result for an aborted request. If
this then gets executed as JS, we get all sorts of errors that we
really don't care about.

This adds the ability to see the aborted state of a request and use
that to return an empty response instead of whatever partial response
is going to be given.

We use the internal `_aborted` flag of the superagent request because
that seems more reliable than other means available from say the XHR.