[FEI-2726].3 Close everything at once
Before, we were nice and went in reverse order, but this meant that the resource loader was the last thing to close. That meant it's 'I'm no longer serving things' code did not get invoked until after the JSDOM window was closed. Now we close everything 'at once' which should have an impact on some of the errors we see that I suspect are down to the shutdown process.

This also makes the virtual console closeable so that we don't get any further logging noise from the JSDOM environment at close time.