[FEI-3495].2 Log request failures
I don't know why we weren't already doing this. We were relying on
the requester to do it, but if it doesn't and we get odd behavior
instead, we can't really debug. So now we log failures. This should
give us some actionable data.

With this in place, if we see partial file results but no error, then
we at least understand a little more of what is not happening.

Test plan:
`yarn test`